requirements.txt
grpcio==1.66.1
grpcio-tools==1.66.1
aiohttp==3.10.5